requirements.txt
sentence_transformers==3.2.1
torch==2.4.1
boto3==1.37.38
requests==2.32.3
pandas==2.0.3
tqdm==4.67.1
flask==1.1.4
Jinja2>=2.11.3,<3.1
MarkupSafe<2.1
faiss-cpu==1.8.0.post1